--- bin/node-template/Cargo.toml
@@ -1,37 +1,0 @@
-[package]
-name = "node-template"
-version = "2.0.0"
-authors = ["Anonymous"]
-build = "build.rs"
-edition = "2018"
-license = "Unlicense"
-
-[[bin]]
-name = "node-template"
-path = "src/main.rs"
-
-[dependencies]
-futures = "0.3.1"
-log = "0.4.8"
-sc-cli = { version = "0.8.0", path = "../../client/cli" }
-sp-core = { version = "2.0.0", path = "../../primitives/core" }
-sc-executor = { version = "0.8", path = "../../client/executor" }
-sc-service = { version = "0.8", path = "../../client/service" }
-sp-inherents = { version = "2.0.0", path = "../../primitives/inherents" }
-sc-transaction-pool = { version = "2.0.0", path = "../../client/transaction-pool" }
-sp-transaction-pool = { version = "2.0.0", path = "../../primitives/transaction-pool" }
-sc-network = { version = "0.8", path = "../../client/network" }
-sc-consensus-aura = { version = "0.8", path = "../../client/consensus/aura" }
-sp-consensus-aura = { version = "0.8", path = "../../primitives/consensus/aura" }
-sp-consensus = { version = "0.8", path = "../../primitives/consensus/common" }
-grandpa = { version = "0.8", package = "sc-finality-grandpa", path = "../../client/finality-grandpa" }
-grandpa-primitives = { version = "2.0.0", package = "sp-finality-grandpa", path = "../../primitives/finality-grandpa" }
-sc-client = { version = "0.8", path = "../../client/" }
-node-template-runtime = { version = "2.0.0", path = "runtime" }
-sp-runtime = { version = "2.0.0", path = "../../primitives/runtime" }
-sc-basic-authorship = { path = "../../client/basic-authorship" }
-structopt = "0.3.8"
-
-[build-dependencies]
-vergen = "3.0.4"
-build-script-utils = { version = "2.0.0", package = "substrate-build-script-utils", path = "../../utils/build-script-utils" }
