+++ bin/node-template/node/src/cli.rs
