+++ bin/node-template/node/src/command.rs
