+++ bin/node-template/pallets/template/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+authors = ['Anonymous']
+edition = '2018'
+name = 'pallet-template'
+version = '2.0.0'
+
+[dependencies]
+codec = { package = "parity-scale-codec", version = "1.0.0", default-features = false, features = ["derive"] }
+safe-mix = { default-features = false, version = '1.0.0' }
+
+[dependencies.frame-support]
+default-features = false
+version = '2.0.0'
+path = "../../../../frame/support"
+
+[dependencies.system]
+default-features = false
+package = 'frame-system'
+version = '2.0.0'
+path = "../../../../frame/system"
+
+[dev-dependencies.sp-core]
+default-features = false
+version = '2.0.0'
+path = "../../../../primitives/core"
+
+[dev-dependencies.sp-io]
+default-features = false
+version = '2.0.0'
+path = "../../../../primitives/io"
+
+[dev-dependencies.sp-runtime]
+default-features = false
+version = '2.0.0'
+path = "../../../../primitives/runtime"
+
+[features]
+default = ['std']
+std = [
+    'codec/std',
+    'frame-support/std',
+    'safe-mix/std',
+    'system/std'
+]
