+++ bin/node-template/runtime/Cargo.toml
@@ -6,6 +6,8 @@ edition = "2018"
 license = "Unlicense"
 
 [dependencies]
+codec = { package = "parity-scale-codec", version = "1.0.0", default-features = false, features = ["derive"] }
+
 aura = { version = "2.0.0", default-features = false, package = "pallet-aura", path = "../../../frame/aura" }
 balances = { version = "2.0.0", default-features = false, package = "pallet-balances", path = "../../../frame/balances" }
 frame-support = { version = "2.0.0", default-features = false, path = "../../../frame/support" }
@@ -16,8 +18,6 @@ sudo = { version = "2.0.0", default-features = false, package = "pallet-sudo", p
 system = { version = "2.0.0", default-features = false, package = "frame-system", path = "../../../frame/system" }
 timestamp = { version = "2.0.0", default-features = false, package = "pallet-timestamp", path = "../../../frame/timestamp" }
 transaction-payment = { version = "2.0.0", default-features = false, package = "pallet-transaction-payment", path = "../../../frame/transaction-payment" }
-
-codec = { package = "parity-scale-codec", version = "1.0.0", default-features = false, features = ["derive"] }
 frame-executive = { version = "2.0.0", default-features = false, path = "../../../frame/executive" }
 serde = { version = "1.0.101", optional = true, features = ["derive"] }
 sp-api = { version = "2.0.0", default-features = false, path = "../../../primitives/api" }
@@ -32,6 +32,8 @@ sp-session = { version = "2.0.0", default-features = false, path = "../../../pri
 sp-std = { version = "2.0.0", default-features = false, path = "../../../primitives/std" }
 sp-transaction-pool = { version = "2.0.0", default-features = false, path = "../../../primitives/transaction-pool" }
 sp-version = { version = "2.0.0", default-features = false, path = "../../../primitives/version" }
+
+template = { version = "2.0.0", default-features = false, path = "../pallets/template", package = "pallet-template" }
 
 [build-dependencies]
 wasm-builder-runner = { version = "1.0.4", package = "substrate-wasm-builder-runner", path = "../../../utils/wasm-builder-runner" }
@@ -64,4 +66,5 @@ std = [
 	"system/std",
 	"timestamp/std",
 	"transaction-payment/std",
+	"template/std",
 ]
